Added Weather + more + bug fixes
I added my weather.

Fixed a bug where the Minutes Of Code would not load. I fixed this by manually changing the minutes and hours.

Moved the time to the top right of the card for a cleaner view.

Added nineteen more quotes.

Made it so that if there's too many to-do list items, the to do list section will scroll.